% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_AFT_comb.R
\name{rankAFT}
\alias{rankAFT}
\title{Parameter Estimation for Rank Based Accelerated Failure Time Model}
\usage{
rankAFT(x.mat, surv.time, surv.cens, method = 1)
}
\arguments{
\item{x.mat}{design matrix containing parameters of interest}

\item{surv.time}{survival time}

\item{surv.cens}{survival censoring variable where 1 represents a failure, and 0 right censoring}
}
\value{
a list containing
 \item{est}{a vector of parameter estimates}
 \item{var}{the individual variance estimate or covariance matrix}
 \item{opt.val}{the initial convergence tolerance (should be near zero)}
}
\description{
This function can be used to estimate the parameters associated with the rank accelerated failure time model as described by Heller (2007)
}
\examples{

n.pts <- 100

betas <- c(-0.2, 0.5, 1)

x.mat <- cbind(rnorm(n.pts, mean = -2, sd = 1),
               rnorm(n.pts, mean = 1, sd = 3),
               rbinom(n.pts, 1, 0.5))

cens.time <- log(runif(n = n.pts, min = 0, max = 50))

eps.i <- rnorm(n = n.pts, mean = 0, sd = 1)

fail.time <- x.mat \%*\% betas + eps.i

surv.time <- exp(pmin(cens.time, fail.time))

surv.cens <- 1 * (fail.time < cens.time)

rankAFT(x.mat, surv.time, surv.cens)

}
\references{
Heller G. (2007) Smoothed Rank Regression With Censored Data
            Journal of the American Statistical Association, 102 (478), 552 - 559.
}
